Kaneki is best girl add Tokyo Ghoul
Skip to content

    About
    Analysis
    Articles
    Awards
    Events
    Fun
    The Regression Sessions

May 2, 2015
The Racial Politics of Tokyo Ghoul

Let’s assume that you and I are human. If we lived in the world of Tokyo Ghoul, it would be our instinct to fear the creatures that lurk in the shadows. If we were investigators, it would be our duty to seek out and destroy the beasts that look like us, but are anything but. These are monsters that kill, who live different lives, have unfamiliar physiologies: They are not like us.

In another blog post, “The Human Heart of Tokyo Ghoul, or Frankenstein’s Monster,” I will touch on the specifics of Tokyo Ghoul’s juxtaposition of monstrosity and humanity. For now, I want to discuss the implications of how Tokyo Ghoul’s portrayal of humans and their relationships with ghouls parallels a very real-world issue of racial politics.

Basically, humans discriminate against ghouls; ghouls discriminate against humans.

    To discriminate is to make a distinction in favour of or against a person or thing on the basis of the group, class, or category to which the person or thing belongs rather than according to actual merit.

yoshi_guy_7681

So in the case of Shinohara who respects the gentle nature, and coffee-making prowess of Yoshimura, there can be no doubt that this ‘thing’ is also part of a category that Shinohara must eradicate, irrespective of any reprievable qualities he might possess.

The ghouls’ seemingly human appearance and their integration into society through Anteiku: Touka’s integration into a school life, Nishiki’s redemption by finding employment in Anteiku, is not enough to vindicate them of their actual crime, that of being a ghoul.

So what about Kaneki Ken?

tokyoghoul-twins

Kaneki was not born a ghoul; he was made a ghoul, as are Kurona and Nashiro Yasuhisa. They are artificially created one-eyed ghouls, so where does that leave them? Kaneki is trapped in the middle, which forces the show to address its racial politics – Kaneki finds himself forced to reconcile his two opposing identities, the human-side and the ghoul-side.

As someone who faces a similar scenario in real life, only without the kagune, I recognise Kaneki’s struggle as being commensurate with a real life crisis of identity.

Throughout Tokyo Ghoul and Root A, and with the manga as well, it is clear that one of the show’s narrative thrusts is to explore Kaneki’s trouble with his (racial) identity.

This struggle is played out from the moment Kaneki becomes a ghoul, where he does not understand why he cannot eat food, and is brought to the fore, when at the end of the first episode, Kaneki is on his knees in front of Touka. The ‘human’ residing in Kaneki battles with the ‘ghoul,’ characterised by Touka who acts as the embodiment of Kaneki’s ‘monstrous’ side:
Tokyo Ghoul Kaneki Human
While Kaneki cannot help but lust after the smell of blood, his human-side is disgusted by the prospect
Tokyo Ghoul Kaneki Touka Eat Blood Hand Flesh
Touka on the other hand, as a full-blooded ghoul, is not concerned, and offers to sate Kaneki’s obvious appetite by forcing him to eat a hand
Tokyo Ghoul Kaneki Humanity
By partaking in this monstrous fruit, Kaneki is scared that he will lose his humanity
Tokyo Ghoul Kaneki Touka Knees
Kaneki is at Touka’s knees, symbolically displaying the strength of the ghoul-side, over the human-side in this moment

So, while Tokyo Ghoul sees Kaneki coming to terms with his identity. Root A sees a seemingly placated Kaneki slowly become undone, as the horror of the reality of his alternative world come crashing against his humanity.

How poignant was the final scenes of Root A, where Eyepatch carries a dead Hide through the army of CCG operatives? How enigmatic was the standoff between Kaneki and the ‘inhuman’ Arima? And that musical backdrop? (The haunting song called ‘Unravel.’)

tokyoghoulmado

Let us talk about the humans in Tokyo Ghoul. Investigator Kureo Mudo is highly respected by his CCG colleagues, but the way he is rendered by the show’s artists gives him a psychotic presence. Arima is heartless by any standard. He lacks emotions and language: two things that evolutionists would say marks out humanity.

Amon and Akira Mado are different. Akira struggles with the loss of her father, admitting to Amon, her grievances regarding what happened on the night her father was killed. This was a very human response – her father being revealed to be the reason she became a ghoul investigator in the first place.

Amon, too, has human reasons for choosing his profession. The man who ran the childrens’ home where Amon grew up was a ghoul. Although Amon survived, it is no surprise that he grows up believing all ghouls are just as monstrous as Priest Porpora. Then he meets Kaneki, who does not fit into this schema: Amon cannot reconcile his notion of a ghoul as a mindless killer to the one who is visibly afraid, and who lets him go at the end of Tokyo Ghoul.

Kaneki_vs_Amon

There are two showdowns in Root A, where Amon desperately tries to find answers that will make sense of the situation; to reconcile his prejudices against ghouls with the notion that they might be sentient beings, capable of judgements.

tokyoghoulanteiku

So, as all good ghouls, we end up in Anteiku. Yoshimura also has a history. He was a ‘monster’ by his own admission, but then he fell in love with a human. The actions of V tore apart that happy home, not the fact that Ukina learned that her husband was a ghoul. This led to the creation of the One-Eyed Owl, a monster that hates the world, protected by a father who is trying to atone for his past mistakes. The One-Eyed Owl may never have existed in her current destructive form, if not, for the actions of V – like Victor Frankenstein, they have, in a sense, created their own monster.

As the series progresses, what we see is an exploration, and various challenges to the assumptions and prejudices that the characters, and the viewers, hold in regards to humans and ghouls – this narrative trajectory exists in the manga, and hopefully will continue in the anime.

In this way, Tokyo Ghoul serves as an allegory of discrimination and racial prejudice, which is challenged by the realisation that fundamentally, we are all sentient. As such, there is such a thing as good and bad in everyone.

Yoshimura typifies this notion, as he acknowledges his sins, his monstrosity, and seeks to change.

The thing I have learned from Tokyo Ghoul is that everyone is capable of change, and that nobody is inherently evil. Just as ghouls are capable of being human, humans are also capable of being monsters.

kaneki-crying